short: Sloos.Pump is no longer used
Long: Sloos.Pump was an attempt to reverse engineer and provide an API
for the way EF reverse engineers a database a creates an assembly.  It
helped me to work through how EF did it job and codify it with code.
EF6 is not supported on .NET Core, so all of my work requires a new
port.  I am not sure there is any value in that because I never really
utilized this code.

I am checkpointing the code as of right now, and my efforts to port it
to .NET Core.  I should (will) delete the code in the future.